TIC-104: Fix import error
